requirements.txt
tensorboard
numpy
scipy
pandas
matplotlib
tqdm
torchinfo
sklearn
seaborn
